move com.google.fonts/check/040 to specs/googlefonts/py
@davelab6 said: "I agree that Lib/fontbakery/specifications/os2.py#L199 looks like a "GF distributor level" check, and so should be in Lib/fontbakery/specifications/googlefonts.py"

https://github.com/googlefonts/fontbakery/issues/1881#issuecomment-392145721

(issue #1881)
